services/auth-service/requirements.txt
Flask==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2
SQLAlchemy==2.0.27
python-dotenv==1.0.0
gunicorn==21.2.0